+++ rails/client/app/bundles/kaiju/components/Component/reducers/reducers.js
@@ -1,10 +1,12 @@
 import { combineReducers } from 'redux';
 import components from './components';
 import selectedComponent from './selectedComponent';
+import refreshedComponent from './refreshedComponent';
 
 const reducers = combineReducers({
   components,
   selectedComponent,
+  refreshedComponent,
 });
 
 export default reducers;